Major refactoring of code base
Separate into dedicated funtions under view
