+++ src/main/python/Views/QELineManager.py
@@ -1,0 +1,111 @@
+import sys
+sys.path.append('..')
+
+from _tools import ProgressBarDecoratorArguments
+
+from os import path
+from PyQt5 import QtWidgets
+import numpy as np
+
+
+@ProgressBarDecoratorArguments(runningText='Generating QELine plot',completedText='QELine plot generated')
+def QELine_plot_button_function(self):    
+    # First check if we have data, otherwise convert data
+    if not self.stateMachine.requireStateByName('Converted'):
+        return False
+    
+    ds = self.DataSetModel.getCurrentDataSet()
+    if len(ds.convertedFiles)==0:
+        self.DataSet_convertData_button_function()
+        
+    # Get the Q points
+    HStart=float(self.ui.QELine_HStart_lineEdit.text())
+    KStart=float(self.ui.QELine_KStart_lineEdit.text())
+    LStart=float(self.ui.QELine_LStart_lineEdit.text())
+    
+    HEnd=float(self.ui.QELine_HEnd_lineEdit.text())
+    KEnd=float(self.ui.QELine_KEnd_lineEdit.text())
+    LEnd=float(self.ui.QELine_LEnd_lineEdit.text())
+            
+    Q1 = np.array([HStart,KStart,LStart])
+    Q2 = np.array([HEnd,KEnd,LEnd])
+    # Collect them into one array
+    QPoints = np.array([Q1,Q2])
+    
+    # Define orthogonal width and minimum pixel size along Q-cut
+    width = float(self.ui.QELine_Width_lineEdit.text())   
+    minPixel = float(self.ui.QELine_MinPixel_lineEdit.text()) 
+
+    # Define energy bins
+    EMin=float(self.ui.QELine_EMin_lineEdit.text())
+    EMax=float(self.ui.QELine_EMax_lineEdit.text())
+    NPoints=int(self.ui.QELine_NPoints_lineEdit.text())        
+    EnergyBins=np.linspace(EMin,EMax,NPoints)
+
+    # Check various plot settings
+    if self.ui.QELine_SelectUnits_RLU_radioButton.isChecked():
+        rlu=True
+    else:
+        rlu=False        
+
+    if self.ui.QELine_LogScale_checkBox.isChecked():
+        log=True
+    else:
+        log=False  
+        
+    if self.ui.QELine_ConstantBins_checkBox.isChecked():
+        constantBins=True
+    else:
+        constantBins=False              
+
+    # Make plot
+    ax,DataLists,Bins,BinCenters,Offsets = \
+    ds.plotCutQELine(QPoints=QPoints, width=width, \
+                                    minPixel=minPixel, EnergyBins=EnergyBins,\
+                                        rlu=rlu,log=log,constantBins=constantBins)
+
+    # Make some final changes to the plot
+    self.QELine=ax    
+    fig = self.QELine.get_figure()
+    fig.set_size_inches(8,6)
+
+    if self.ui.QELine_Grid_checkBox.isChecked():
+        ax.grid(True)
+    else:
+        ax.grid(False)
+    
+    self.QELine_setCAxis_button_function()
+
+    self.windows.append(self.QELine.get_figure())
+    self.QELine_SetTitle_button_function()
+
+    return True
+
+
+def QELine_setCAxis_button_function(self):       
+    CAxisMin=float(self.ui.QELine_CAxisMin_lineEdit.text())
+    CAxisMax=float(self.ui.QELine_CAxisMax_lineEdit.text())
+    
+    self.QELine.set_clim(CAxisMin,CAxisMax)
+    fig = self.QELine.get_figure()
+    fig.canvas.draw()
+
+def QELine_SetTitle_button_function(self):
+    if hasattr(self, 'QELine'):
+        TitleText=self.ui.QELine_SetTitle_lineEdit.text()        
+        self.QELine.set_title(TitleText)
+        fig = self.QELine.get_figure()
+        fig.canvas.draw()
+
+
+
+def initQELineManager(guiWindow):
+    guiWindow.QELine_plot_button_function = lambda: QELine_plot_button_function(guiWindow)
+    guiWindow.QELine_setCAxis_button_function = lambda: QELine_setCAxis_button_function(guiWindow)
+    guiWindow.QELine_SetTitle_button_function = lambda: QELine_SetTitle_button_function(guiWindow)
+
+def setupQELineManager(guiWindow):
+    
+    guiWindow.ui.QELine_plot_button.clicked.connect(guiWindow.QELine_plot_button_function)
+    guiWindow.ui.QELine_setCAxis_button.clicked.connect(guiWindow.QELine_setCAxis_button_function)
+    guiWindow.ui.QELine_SetTitle_button.clicked.connect(guiWindow.QELine_SetTitle_button_function)
